Introduce AddressUtil
