+++ test/config/ConfigBuilderTest.js
@@ -25,7 +25,10 @@ const {
     TopicOverloadPolicy
 } = require('../..');
 const { ConfigBuilder } = require('../../lib/config/ConfigBuilder');
-const { AddressHelper } = require('../../lib/util/Util');
+const {
+    getSocketAddresses,
+    createAddressFromString
+} = require('../../lib/util/AddressUtil');
 const { ReconnectMode } = require('../../lib/config/ConnectionStrategyConfig');
 
 describe('ConfigBuilderTest', function () {
@@ -98,7 +101,7 @@ describe('ConfigBuilderTest', function () {
     it('network', function () {
         const networkCfg = fullConfig.network;
 
-        const addresses0 = AddressHelper.getSocketAddresses(networkCfg.clusterMembers[0]);
+        const addresses0 = getSocketAddresses(networkCfg.clusterMembers[0]);
         expect(addresses0[0].host).to.equal('127.0.0.9');
         expect(addresses0[0].port).to.equal(5701);
         expect(addresses0[1].host).to.equal('127.0.0.9');
@@ -107,13 +110,13 @@ describe('ConfigBuilderTest', function () {
         expect(addresses0[2].port).to.equal(5703);
         expect(addresses0.length).to.equal(3);
 
-        const addresses1 = AddressHelper.getSocketAddresses(networkCfg.clusterMembers[1]);
+        const addresses1 = getSocketAddresses(networkCfg.clusterMembers[1]);
         expect(addresses1[0].host).to.equal('127.0.0.2');
         expect(addresses1[0].port).to.equal(5702);
         expect(addresses1.length).to.equal(1);
 
-        const address0 = AddressHelper.createAddressFromString(networkCfg.clusterMembers[0]);
-        const address1 = AddressHelper.createAddressFromString(networkCfg.clusterMembers[1]);
+        const address0 = createAddressFromString(networkCfg.clusterMembers[0]);
+        const address1 = createAddressFromString(networkCfg.clusterMembers[1]);
         expect(address0.host).to.equal('127.0.0.9');
         expect(address0.port).to.be.undefined;
         expect(address1.host).to.equal('127.0.0.2');
